begin on custom, cached array roots merkleization
