+++ shared/stateutil/BUILD.bazel
@@ -3,6 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "go_default_library",
     srcs = [
+        "arrays.go",
         "attestations.go",
         "blocks.go",
         "helpers.go",
